+++ src/models/Entry.php
@@ -478,7 +478,8 @@ class Entry extends ActiveRecord implements AssetParentInterface
 
     public function hasRoute(): bool
     {
-        return $this->section_count > 0;
+        return $this->section_count > 0
+            || (in_array('parent_slug', $this->slugTargetAttribute) && $this->entry_count > 0);
     }
 
     public function isIndex(): bool